requirements.txt
Django==2.0.6
django-cors-headers==2.2.0
django-tastypie==0.14.1
neo4j-driver==1.6.0
neotime==1.0.0
python-dateutil==2.7.3
python-mimeparse==1.6.0
pytz==2018.4
six==1.11.0
